This adds a lot of USB capture data to ensure that we understand the order of opreation required to use the reader. I have commented on what I was doing when I pushed data in and out of the USB bus. I have also included the card that I used, should anyone want to try to take a stab at comparing the bytes from the usb bus to the bytes in the MAG file.
This presents a lot of work but should include everything we need for a basic makstripe.c implementation.

It appears that everything not included in this dump is either not implemented in the linux gui software or it is done in that software. All signal processing is clearly done in the software and not in the MAKStripe. It's not unlike a soft modem. This fits with the understanding of the device, the size of the device and of course the heavy duty software required even for an ISO read or ISO write.